removed warnings

+++ de.unikn.knime.core/src/de/unikn/knime/core/data/util/ObjectToDataCellConverter.java
@@ -163,7 +163,7 @@ public class ObjectToDataCellConverter {
      * @see IntCell
      */
     public DataCell createDataCell(final byte b) {
-        return new IntCell((int)b);
+        return new IntCell(b);
     }
 
     /**
